+++ api/.gitignore
@@ -3,3 +3,4 @@ node_modules
 *.sql
 .env
 dist/
+*.log
