+++ api/package.json
@@ -10,18 +10,18 @@
     "jsdom": "^21.1.1",
     "mysql": "^2.18.1",
     "mysql2": "^3.3.1",
-    "nodemon": "^2.0.22"
+    "nodemon": "^2.0.22",
+    "typescript": "^5.1.3"
   },
   "scripts": {
-    "start": "node index.js",
+    "start": "node dist/index.js",
     "dev": "nodemon -L dist/index.js",
-    "build": "babel src -d dist"
+    "build": "tsc"
   },
   "devDependencies": {
-    "@babel/cli": "^7.21.5",
-    "@babel/core": "^7.22.1",
-    "@babel/preset-env": "^7.22.4",
+    "@types/cookie-parser": "^1.4.3",
     "@types/express": "^4.17.17",
+    "@types/node": "^20.2.5",
     "core-js": "^3.30.2",
     "regenerator-runtime": "^0.13.11",
     "webpack-cli": "^5.1.3"
