+++ api/src/types/request.ts
@@ -1,0 +1,5 @@
+interface onlineRequestBody{
+  threshold: undefined | number
+}
+
+export { onlineRequestBody }
