+++ api/src/types/table.ts
@@ -1,0 +1,13 @@
+export interface Timeline {
+  timeline_id: number
+  player_name: string
+  ranking: number
+  achievement: string
+  chara: string
+  point: number
+  created_at: string
+  diff: number
+  last_timeline_id: number
+  elapsed: number
+}
+
