+++ api/webpack.config.js
@@ -1,16 +1,25 @@
-
 module.exports = {
   target: "node",
   entry: "./src/index.js",
   module: {
     rules: [
       {
-        test: /\.m?js$/,
+        test: /\.(js|ts)x?$/,
         exclude: /node_modules/,
         use: {
-          loader: "babel-loader",
-        },
-      },
+          loader: 'babel-loader',
+          options: {
+            presets: [
+              '@babel/preset-env',
+              '@babel/preset-typescript',
+              '@babel/preset-react'
+            ],
+            plugins: [
+              '@babel/plugin-proposal-class-properties'
+            ]
+          }
+        }
+      }
     ],
   },
 };
